refactor(dashboard): remove unused dashboard project references and clean up code
