+++ src/Atomizer.Dashboard/Endpoints/QueueStatsEndpoints.cs
@@ -1,3 +1,4 @@
+using Atomizer.Abstractions;
 using Atomizer.Dashboard.Contracts;
 using Microsoft.AspNetCore.Http;
 using Microsoft.Extensions.DependencyInjection;
@@ -8,7 +9,7 @@ internal static class QueueStatsEndpoints
 {
     internal static async Task GetStatsAsync(HttpContext context)
     {
-        var storage = context.RequestServices.GetRequiredService<IAtomizerDashboardStorage>();
+        var storage = context.RequestServices.GetRequiredService<IAtomizerStorage>();
         var stats = await storage.GetQueueStatsAsync(context.RequestAborted);
         var response = new QueueStatsResponse { Queues = stats.Select(QueueStatsDto.From).ToList() };
         await DashboardJsonResponse.WriteAsync(context, response, context.RequestAborted);
