+++ src/Atomizer.Dashboard/Endpoints/ServersEndpoints.cs
@@ -1,3 +1,4 @@
+using Atomizer.Abstractions;
 using Atomizer.Core;
 using Atomizer.Dashboard.Contracts;
 using Microsoft.AspNetCore.Http;
@@ -9,7 +10,7 @@ internal static class ServersEndpoints
 {
     internal static async Task ListAsync(HttpContext context)
     {
-        var storage = context.RequestServices.GetRequiredService<IAtomizerDashboardStorage>();
+        var storage = context.RequestServices.GetRequiredService<IAtomizerStorage>();
         var clock = context.RequestServices.GetRequiredService<IAtomizerClock>();
         var servers = await storage.GetActiveServersAsync(context.RequestAborted);
         var now = clock.UtcNow;
