+++ src/Atomizer.EntityFrameworkCore/Atomizer.EntityFrameworkCore.csproj
@@ -22,6 +22,5 @@
     </ItemGroup>
     <ItemGroup>
         <ProjectReference Include="..\Atomizer\Atomizer.csproj" />
-        <ProjectReference Include="..\Atomizer.Dashboard\Atomizer.Dashboard.csproj" />
     </ItemGroup>
 </Project>
