+++ src/Atomizer/Atomizer.csproj
@@ -17,9 +17,6 @@
     </PropertyGroup>
     <ItemGroup>
         <AssemblyAttribute Include="System.Runtime.CompilerServices.InternalsVisibleToAttribute">
-            <_Parameter1>Atomizer.Dashboard</_Parameter1>
-        </AssemblyAttribute>
-        <AssemblyAttribute Include="System.Runtime.CompilerServices.InternalsVisibleToAttribute">
             <_Parameter1>Atomizer.Tests</_Parameter1>
         </AssemblyAttribute>
     </ItemGroup>
