+++ src/Atomizer/Storage/InMemoryStorage.cs
@@ -374,6 +374,99 @@ public sealed class InMemoryStorage : IAtomizerStorage
     }
 
     /// <inheritdoc/>
+    public Task<PagedResult<AtomizerJob>> GetJobsAsync(JobQuery query, CancellationToken cancellationToken)
+    {
+        cancellationToken.ThrowIfCancellationRequested();
+
+        var take = Math.Min(query.Take, 500);
+
+        IEnumerable<AtomizerJob> jobs = _jobs.Values;
+
+        if (query.Statuses is { Count: > 0 })
+            jobs = jobs.Where(j => query.Statuses.Contains(j.Status));
+
+        if (query.QueueKey is not null)
+            jobs = jobs.Where(j => j.QueueKey == query.QueueKey);
+
+        if (query.PayloadTypeName is not null)
+            jobs = jobs.Where(j =>
+                j.PayloadType != null
+                && j.PayloadType.Name.IndexOf(query.PayloadTypeName, StringComparison.OrdinalIgnoreCase) >= 0
+            );
+
+        if (query.CreatedFromUtc.HasValue)
+            jobs = jobs.Where(j => j.CreatedAt >= query.CreatedFromUtc.Value);
+
+        if (query.CreatedToUtc.HasValue)
+            jobs = jobs.Where(j => j.CreatedAt <= query.CreatedToUtc.Value);
+
+        var ordered = jobs.OrderByDescending(j => j.CreatedAt).ToList();
+        var total = ordered.Count;
+        var items = ordered.Skip(query.Skip).Take(take).ToList();
+
+        return Task.FromResult(
+            new PagedResult<AtomizerJob>
+            {
+                Items = items,
+                TotalCount = total,
+                Skip = query.Skip,
+                Take = take,
+            }
+        );
+    }
+
+    /// <inheritdoc/>
+    public Task<AtomizerJob?> GetJobByIdAsync(Guid id, CancellationToken cancellationToken)
+    {
+        cancellationToken.ThrowIfCancellationRequested();
+
+        _jobs.TryGetValue(id, out var job);
+        return Task.FromResult<AtomizerJob?>(job);
+    }
+
+    /// <inheritdoc/>
+    public Task<IReadOnlyList<AtomizerSchedule>> GetSchedulesAsync(CancellationToken cancellationToken)
+    {
+        cancellationToken.ThrowIfCancellationRequested();
+
+        lock (_syncRoot)
+        {
+            return Task.FromResult((IReadOnlyList<AtomizerSchedule>)_schedules.Values.ToList());
+        }
+    }
+
+    /// <inheritdoc/>
+    public Task<IReadOnlyList<AtomizerActiveServer>> GetActiveServersAsync(CancellationToken cancellationToken)
+    {
+        cancellationToken.ThrowIfCancellationRequested();
+
+        lock (_syncRoot)
+        {
+            return Task.FromResult((IReadOnlyList<AtomizerActiveServer>)_activeServers.Values.ToList());
+        }
+    }
+
+    /// <inheritdoc/>
+    public Task<IReadOnlyList<QueueStats>> GetQueueStatsAsync(CancellationToken cancellationToken)
+    {
+        cancellationToken.ThrowIfCancellationRequested();
+
+        var groups = _jobs
+            .Values.GroupBy(j => j.QueueKey)
+            .Select(g => new QueueStats
+            {
+                QueueKey = g.Key,
+                Pending = g.Count(j => j.Status == AtomizerJobStatus.Pending),
+                Processing = g.Count(j => j.Status == AtomizerJobStatus.Processing),
+                Completed = g.Count(j => j.Status == AtomizerJobStatus.Completed),
+                Failed = g.Count(j => j.Status == AtomizerJobStatus.Failed),
+            })
+            .ToList();
+
+        return Task.FromResult<IReadOnlyList<QueueStats>>(groups);
+    }
+
+    /// <inheritdoc/>
     public async Task<TResult> ExecuteInLeaseAsync<TResult>(
         QueueKey queue,
         Func<CancellationToken, Task<TResult>> callback,
@@ -416,26 +509,6 @@ public sealed class InMemoryStorage : IAtomizerStorage
             },
             cancellationToken
         );
-
-    // ---- dashboard snapshot accessors ----
-
-    internal IReadOnlyList<AtomizerJob> GetAllJobs() => _jobs.Values.ToList();
-
-    internal IReadOnlyList<AtomizerSchedule> GetAllSchedules()
-    {
-        lock (_syncRoot)
-        {
-            return _schedules.Values.ToList();
-        }
-    }
-
-    internal IReadOnlyList<AtomizerActiveServer> GetAllServers()
-    {
-        lock (_syncRoot)
-        {
-            return _activeServers.Values.ToList();
-        }
-    }
 
     // ---- helpers ----
