+++ tests/Atomizer.EntityFrameworkCore.Tests/Dashboard/SqlServerStorageMonitoringTests.cs
@@ -1,15 +1,15 @@
 using Atomizer.EntityFrameworkCore.Tests.Fixtures;
 using Atomizer.EntityFrameworkCore.Tests.Storage;
-using Atomizer.EntityFrameworkCore.Tests.TestSetup.MySql;
+using Atomizer.EntityFrameworkCore.Tests.TestSetup.SqlServer;
 using AwesomeAssertions;
 
 namespace Atomizer.EntityFrameworkCore.Tests.Dashboard;
 
-[Collection(nameof(MySqlDatabaseFixture))]
-public sealed class MySqlDashboardStorageTests(MySqlDatabaseFixture fixture)
-    : EntityFrameworkCoreDashboardStorageTests<MySqlDbContext>
+[Collection(nameof(SqlServerDatabaseFixture))]
+public sealed class SqlServerStorageMonitoringTests(SqlServerDatabaseFixture fixture)
+    : EntityFrameworkCoreStorageMonitoringTests<SqlServerDbContext>
 {
-    protected override MySqlDbContext CreateDbContext() => fixture.CreateNewDbContext();
+    protected override SqlServerDbContext CreateDbContext() => fixture.CreateNewDbContext();
 
     public override ValueTask InitializeAsync() => ValueTask.CompletedTask;
 
@@ -21,7 +21,7 @@ public sealed class MySqlDashboardStorageTests(MySqlDatabaseFixture fixture)
     }
 
     [Fact]
-    public async Task GetJobsAsync_WhenUsingMySql_ShouldReturnPagedResults()
+    public async Task GetJobsAsync_WhenUsingMsSql_ShouldReturnPagedResults()
     {
         var result = await CreateStorage().GetJobsAsync(new JobQuery { Skip = 0, Take = 10 }, CancellationToken.None);
