+++ tests/Atomizer.Tests/Atomizer.Tests.csproj
@@ -38,6 +38,5 @@
     </ItemGroup>
     <ItemGroup>
         <ProjectReference Include="..\Atomizer.Tests.Utilities\Atomizer.Tests.Utilities.csproj" />
-        <ProjectReference Include="..\..\src\Atomizer.Dashboard\Atomizer.Dashboard.csproj" />
     </ItemGroup>
 </Project>
